[backend] update the route

+++ Backend/Andromeda/routes/api.php
@@ -15,20 +15,20 @@ Route::group([
     // Route::get('notifications', 'AuthController@notifications');
     
     //-------------------------USER---------------------------
-    Route::get('user/showall','UserController@index');//* display all Users
-    Route::get('user/{id}/show','UserController@show');//* Display the specified User
-    Route::post('user/{user}/update','UserController@update');
-    Route::delete('user/{user}/delete','UserController@destroy');
+    Route::get('user','UserController@index');//* display all Users
+    Route::get('user/{user}','UserController@show');//* Display the specified User
+    Route::post('user/{user}/update','UserController@update');//* update the specified User
+    Route::delete('user/{user}/delete','UserController@destroy');//* delete the specified User
 
     //-------------------------COURSE---------------------------
-    Route::get('course/showall','CourseController@index'); //* display all courses
-    Route::get('course/{id}/show','CourseController@show');//* display specific course
-    Route::get('course/progression','CourseController@courseProgressions');//* display specific course
+    Route::get('course','CourseController@index'); //* display all courses
+    Route::get('course/progression','CourseController@courseProgressions');//* display all courses followed with progression
+    Route::get('course/{course}','CourseController@show');//* display specific course
 
     //-------------------------NOTIFICATION---------------------------
-    Route::get('notification/showall','NotificationController@index'); //* display all notification belongs to currently authenticated user
+    Route::get('notification','NotificationController@index'); //* display all notification belongs to currently authenticated user
     Route::get('notification/{notification}','NotificationController@show');//* display specific notification belongs to currently authenticated user
-    Route::delete('notification/{notification}','NotificationController@destroy');
+    Route::delete('notification/{notification}','NotificationController@destroy');//* delete specific notification belongs to currently authenticated user
 
     //-------------------------DISCUSSION--------------------------------
     Route::get('discussion/','DiscussionController@index');//* display all discussions belongs to currently authenticated user
